+++ lib/puppet_x/encore/powercli/helper.rb
@@ -1,4 +1,5 @@
 require 'puppet_x'
+require 'ruby-pwsh'
 require 'singleton'
 
 module PuppetX::PowerCLI